requirements.txt
discord.py
python-dotenv
APScheduler
psycopg2-binary
google-genai